// src/types/entities/services/Sbom.ts
export enum Status {
    Success = 'success',
    Failure = 'failure'
}

export interface Dependency {
    Key: string;
    Requires: { [key: string]: string };
    Dependencies: { [key: string]: string };
    Optional: boolean;
    Bundled: boolean;
    Dev: boolean;
    Transitive: boolean;
    Licenses: string[];
}
// ...
interface WorkSpaceDependency {
    name: string;
    version: string;
    constraint: string;
}

export interface WorkSpaceData {
    dependencies: { [key: string]: { [key: string]: Dependency } };
    start: {
        dependencies?: WorkSpaceDependency[];
        dev_dependencies?: WorkSpaceDependency[];
    };
}

export interface Output {
    workspaces: { [key: string]: WorkSpaceData };
    analysis_info: AnalysisInfo;
}

export interface AnalysisInfo {
    status: Status;
    private_errors: StatusError[];
    public_errors: StatusError[];
    project_name: string;
    working_directory: string;
    package_manager: string;
    lock_file_version: number;
    lock_file_path: string;
    package_file_path: string;
    relative_lock_file_path: string;
    relative_package_file_path: string;
    analysis_start_time: string;
    analysis_end_time: string;
    analysis_delta_time: number;
    version_seperator: string;
    import_path_seperator: string;
    default_workspace_name: string;
    self_managed_workspace_name: string;
    work_spaces_used: boolean;
    work_space_package_file_paths: { [key: string]: string };
    stats: { [key: string]: any };
}

export interface LicenseDist {
    [licenseId: string]: number;
}

export interface Stats {
    license_dist: LicenseDist;
    number_of_spdx_licenses: number;
    number_of_non_spdx_licenses: number;
    number_of_copy_left_licenses: number;
    number_of_permissive_licenses: number;
}

export interface StatusError {
    type: string;
    description: string;
}